Minor changes to the log levels
For setter and getter, log level depends on log level of the object
which we send as an argument. If we do not specify this level for the
each object which we send to setter, Tools method uses the default value
in the Base class.
Also in setter and getter, Physics and Geometry use log level of the
Fluid and the Network respectively, (not their own levels).


Former-commit-id: 088eaa6c56d935ca806a9cb9d6a004855a1bcfb0
Former-commit-id: 246bc8f1c0ae5d241fe822f16c7cd7a1ce09084a