Minor changes to Linear Solver.
Now, even if the user wants to use  a transport algorithm (with known
BC) to calculate effective permeability and diffusivity, the Linear
Solver keeps the user BC and does not overwrite it during its
calculation.
However, it is still possible to calculate these values without running
a  transport algorithm separately. User should just create a transport
algorithm object and then the method will be available. In this case,
there is no need to apply BC, because the default boundary conditions
are embedded in Linear Solver.