Linear solver has become faster than before.
By changing the logic for adding diagonal values of Non-Dirichlet pores,
it has become faster. The slowest part is still
scipy.sparse.linalg.spsolve. The default argument of this solver is
use_umfpack=True, which makes the solver fast for sparse matrices. But
remember that for Np>300k,, if you use this flag,  you may get an error
due to insufficient memory for  'umfpack'.


Former-commit-id: 3ca9981699ebcbc69cc9da27c8612bab74f31f90
Former-commit-id: d82b46b32e9fe4aa0129348a6f1cea34f1693ac5