Some changes to BC to make it more user friendly.
Now, it is possible for user to apply a value as Neumann BC to a group
of pores or add a value to each pore as Neumann BC. We had this
functionality before, but right now, it is easier. So, instead of
'Neumann_rate', two new BC have been added:
Neumann_rate_union  and Neumann_rate_individual.


Former-commit-id: ad921edc404c50b1fbc14f147a45bb2072e427da
Former-commit-id: fb525e2325f87dfbd4ea9d5e811e5d760d389d03