Some changes to BC to make it more user friendly.
Now, it is possible for user to apply a value as Neumann BC to a group
of pores or add a value to each pore as Neumann BC. We had this
functionality before, but right now, it is easier. So, instead of
'Neumann_rate', two new BC have been added:
Neumann_rate_union  and Neumann_rate_individual.