debugged subdivide method
This is just a quick fix for the problem that some users might have had.
Before, it was allowing user to run the subdivide method consecutively,
but right now, it raises an Exception. Because creating the proper
connections is really complicated when it comes to the different spacing
and scales.