bulk_diffusion now uses correct diffusivities for pores.
It used to use the interpolated throat diffusivity for pores as well. Now, it uses pore diffusivity for pores and interpolates for throat diffusivity. Also, if 'throat.diffusivity' is present in phase, it doesn't interpolate. A use case is in presence of Knudsen effect where one might calculate diffusivity for pores and throats separately.