change A.dtype to int64
fix Memory limit for scipy.sparse.linalg.spsolve with scikit-umfpack. 
when using umfpack to solve Ax=b, which have 10X speedup in Ubuntu and MacOS，an int32 indices will hit a 32 bit memory limit. This is because the 64-bit umfpack family solver is only called when the indices have dtype int64 (see more: https://github.com/scipy/scipy/issues/8278). An easily patch for this bug is changing the dtype of A to np.int64