local changes of rel_diff algorithm [ci skip]
I removed some explanation and added other explanations. I also wrote the code for creating 6 Fickian diffusion outside the loop and running the algorithms and regenerate models in the loop, but that had a similar run time to the previous approach of creating all the algorithms inside the loop. I removed my changes and presented the cell as it was before.